Some refactoring and add competitionscontroller
- Also move some stuff to friends manager
