+++ Project SF/Logic/Friends/FriendStruct.swift
