+++ Project SF/Views/Tabs/Competitions/CreateCompetition.swift
@@ -26,7 +26,7 @@ struct CreateCompetition: View {
     @State var distanceGoal = "10"
     @State var distanceGoalInt = 10
     
-    let competitionController = CompetitionsController()
+    @EnvironmentObject var competitionController: CompetitionsController
 
     var body: some View {
         ScrollView(showsIndicators: false) {
@@ -204,24 +204,20 @@ struct CreateCompetition: View {
             RoundedButton("Start the competition") {
                 guard !competitionName.isEmpty else { return }
                 
-                competitionController.createCompetition(
-                    type: .init(
-                        move: move,
-                        exercise: exercise,
-                        stand: stand,
-                        steps: steps,
-                        distance: distance,
-                        stepsGoal: stepsGoalInt,
-                        distanceGoal: distanceGoalInt
-                    ),
+                let newCompetition = Competition(
                     title: competitionName,
-                    endDate: competitionEndDate,
-                    // TODO: Add support for inviting friends when creating competition.
-                    friends: [],
-                    then: { result in
-                        print(result)
-                    }
+                    move: move,
+                    exercise: exercise,
+                    stand: stand,
+                    steps: steps,
+                    distance: distance,
+                    stepsGoal: stepsGoalInt,
+                    distanceGoal: distanceGoalInt,
+                    startDate: Date(),
+                    endDate: competitionEndDate
                 )
+                
+                competitionController.create(competition: newCompetition, with: [])
                 
                 presentationMode.wrappedValue.dismiss()
             }
